wealth-calculator/requirements-api.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
snowflake-connector-python==3.12.3
python-dotenv==1.0.1
